TandemBluetoothHandler: automatically unbond when unable to progress past initialPumpConnection
This is an attempt to work around what is presumed to be an Android bug with bonded peripherals
which results in the phone never receiving a reply after an initial connection is established
and the first auth packet is sent to the pump. Manually unbonding the peripheral in Bluetooth
settings seems to resolve this problem when it occurs.